Break fab tasks into separate namespaces
