+++ fabfile/config.py
@@ -1,0 +1,17 @@
+paths = {
+    'tests-repo': '/home/ec2-user/test-automation',
+    'citus-repo': '/home/ec2-user/citus',
+    'enterprise-repo': '/home/ec2-user/citus-enterprise',
+    'session-repo': '/home/ec2-user/session-analytics',
+    'hll-repo': '/home/ec2-user/hll',
+    'pg-latest': '/home/ec2-user/pg-latest',
+    'pg-source-balls': '/home/ec2-user/postgres-source',
+}
+
+settings = {
+    'citus-git-ref': 'master',
+    'session-analytics-ref': 'master',
+    'hll-ref': 'v2.10.0',
+    'pg-version': '9.6.1',
+    'pg-configure-flags': [],
+}
